utils/telegram_utils.py
import requests
import os

from dotenv import load_dotenv

load_dotenv()

token = os.getenv("TELEGRAM_TOKEN")
chat_id = os.getenv("TELEGRAM_ID")
message = ()


def send_telegram_message(message):
    url = f"https://api.telegram.org/bot{token}/sendMessage?chat_id={chat_id}&text={message}&parse_mode=html"
    print(requests.get(url).json())
