requirements.txt
black==23.3.0
openai==0.27.7
praw==7.7.0
prawcore==2.3.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0